move chain DAO to ChainStore.ts
